Added manifest.del functionality to start-merge script. This will synchronize processing of the incremental backup directories with the restore script.
If a .lock file is present in the incremental backup directory, we will not delete the merged
incremental backup files. Instead a timestamp-manifest.del file will be added to that directory.

In the next pass of merge-script, it will check if a *-manifest.del file exists, if it does then
it will delete the list of the files listed in the manifest.del file.

If a .lock file is present along with the manifest.del file then start-merge will pass the manifest.del
file as an -x argument to mbmerge-incr script. The latter should skip the files listed
in the manifest file while doing the merge for that directory. This functionality in mbmerge-incr
is currently broken right now.

On completion of the merge process, start-merge will append the list of files that were processed
in the current pass to the existing manifest-del file and replace the old manifest with the
new one.