SEG-9042 - Fix manifest.del to use append for adding files
When daily-merge is executed and manifest.del is create since
a lock file is present, it will ignore the deletion of source files
from which daily backup is created. When daily-merge attempts the
second time and then also lock file is present, it will overwrite
the manifest.del file with new list of files to be deleted. This
patch fixes the overwrite with append.